enhancement: per allele base coverage counter capped at uint16_t max

+++ libgramtools/include/quasimap/coverage/types.hpp
@@ -13,7 +13,7 @@ using GroupedAlleleCounts = SequenceHashMap<AlleleIds, uint64_t>;
 using SitesGroupedAlleleCounts = std::vector<GroupedAlleleCounts>;
 using AlleleGroupHash = SequenceHashMap<AlleleIds, uint64_t>;
 
-using BaseCoverage = std::vector<uint64_t>;
+using BaseCoverage = std::vector<uint16_t>;
 using AlleleCoverage = std::vector<BaseCoverage>;
 using SitesAlleleBaseCoverage = std::vector<AlleleCoverage>;
